refactor(fe): refactor atom
