--- frontend/src/api/selector.js
@@ -1,1 +1,0 @@
-import { selector, selectorFamily } from "recoil";
