+++ frontend/src/atom/testAtom.js
@@ -1,0 +1,43 @@
+import { atom } from "recoil";
+
+const localStorageEffect =
+  (key) =>
+  ({ setSelf, onSet }) => {
+    const savedValue = localStorage.getItem(key);
+    if (savedValue != null) {
+      setSelf(JSON.parse(savedValue));
+    }
+
+    onSet((newValue) => {
+      if (newValue === null) {
+        localStorage.removeItem(key);
+      } else {
+        localStorage.setItem(key, JSON.stringify(newValue));
+      }
+    });
+  };
+
+export const timeLeftState = atom({
+  key: "timeLeftState",
+  default: 120,
+});
+
+export const isRunningState = atom({
+  key: "isRunningState",
+  default: false,
+});
+
+export const isStoppedState = atom({
+  key: "isStoppedState",
+  default: false,
+});
+
+export const audioURLState = atom({
+  key: "audioURLState",
+  default: null,
+});
+
+export const isRecordingState = atom({
+  key: "isRecordingState",
+  default: false,
+});
