+++ frontend/src/pages/Edit.jsx
@@ -1,10 +1,9 @@
 import { useRecoilState } from "recoil";
-import { userInfoState } from "../api/atom";
+import { userInfoState, userPkState } from "../atom/authAtoms.js";
 import { useNavigate } from "react-router-dom";
 import OnboardingSurvey from "../components/homePage/OnboardingSurvey";
 import { updateUserInfoAPI } from "../api/api";
 import SideBar from "../components/sideBar/SideBar.jsx";
-import { userPkState } from "../api/authAtoms.js";
 
 const Edit = () => {
   const [userData, setUserData] = useRecoilState(userInfoState);
